edit enquiries, videos, catalogues request body and fixing gallery and images uploads

+++ src/controllers/manufacturer/projects/updateProject.ts
@@ -53,7 +53,7 @@ export const updateProject = async (req: Request, res: Response) =>{
                 if(project.value.coverImage){
                     await storage.bucket(`${process.env.GCS_BUCKET}`).file(project.value.coverImage.name).delete();            
                 }
-                if(project.value.images && project.value.images.length>0){
+                if(files[1] && project.value.images && project.value.images.length>0){
                     (project.value.images as Array<IFile>).map(async image =>
                         await storage.bucket(`${process.env.GCS_BUCKET}`).file(image.name).delete())
                     }
